OpenAI-SDK-compatable-API/requirements.txt
fastapi==0.111.0
uvicorn==0.30
openai==1.34.0
tinydb==4.8.0
gunicorn==22.0.0
uvloop==0.19.0
httptools==0.6.1
python-dotenv==1.0.0